DMP-2626 moved existing tests to abstract class
